+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/control/gainmatrices/RGB.kt
@@ -1,7 +1,7 @@
 package org.firstinspires.ftc.teamcode.control.gainmatrices
 
 data class RGB @JvmOverloads constructor(
-    @JvmField var red: Float = 0.0F,
-    @JvmField var green: Float = 0.0F,
-    @JvmField var blue: Float = 0.0F,
+    @JvmField var red: Double = 0.0,
+    @JvmField var green: Double = 0.0,
+    @JvmField var blue: Double = 0.0,
 )
